+++ 24th-App-Team-1-iOS/Domain/CommonDomain/Sources/Repositories/CommonRepositoryProtocol.swift
@@ -13,5 +13,6 @@ public protocol CommonRepositoryProtocol {
     func createCheckProfanity(body: CreateCheckProfanityRequest) -> Single<Bool>
     func fetchProfileImages() -> Single<FetchProfileImageResponseEntity?>
     func fetchProfileBackgrounds() -> Single<FetchProfileBackgroundsResponseEntity?>
+    func createRefreshToken(body: CreateRefreshTokenRequest) -> Single<CreateAccountResponseEntity?> // 토큰 재발행
     func createReportUserItem(body: CreateUserReportRequest) -> Single<CreateReportUserEntity?>
 }
