+++ 24th-App-Team-1-iOS/Domain/LoginDomain/Sources/Repositories/LoginRepositoryProtocol.swift
@@ -13,6 +13,5 @@ public protocol LoginRepositoryProtocol {
     func createNewMemberToken(body: CreateSignUpTokenRequest) -> Single<CreateSignUpTokenResponseEntity?> // 신규회원
     func createExistingMember(body: CreateSignUpTokenRequest) -> Single<CreateAccountResponseEntity?> // 기존회원
     func createAccount(body: CreateAccountRequest) -> Single<CreateAccountResponseEntity?> //회원가입
-    func createRefreshToken(body: CreateRefreshTokenRequest) -> Single<CreateAccountResponseEntity?> // 토큰 재발행
     func fetchSchoolList(query: SchoolListRequestQuery) -> Single<SchoolListResponseEntity?> // 학교 검색
 }
