+++ 24th-App-Team-1-iOS/Service/CommonService/Sources/DataMapping/CreateAccountResponseDTO.swift
@@ -1,0 +1,22 @@
+//
+//  CreateAccountResponseDTO.swift
+//  LoginService
+//
+//  Created by eunseou on 7/30/24.
+//
+
+import Foundation
+import LoginDomain
+
+public struct CreateAccountResponseDTO: Decodable {
+    public let accessToken: String
+    public let refreshToken: String
+    public let refreshTokenExpiredAt: String
+    
+}
+
+extension CreateAccountResponseDTO {
+    func toDomain() -> CreateAccountResponseEntity {
+        return .init(accessToken: accessToken, refreshToken: refreshToken, refreshTokenExpiredAt: refreshTokenExpiredAt)
+    }
+}
